test: isolate qwen output bias B200 coverage
